Move componetns outside App router
