+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import SearchBar from "./components/SearchBar";
+import SearchBar from "../components/SearchBar";
 
 export default function MainPage() {
   return (
